requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
Django==3.0.3
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-environ==0.4.5
django-js-asset==1.2.2
importlib-metadata==1.5.0
kombu==4.6.7
pytz==2019.3
sqlparse==0.3.0
vine==1.3.0
zipp==3.0.0
psycopg2-binary
